Fix many_till localy.
Include a local copy of the many_till macro from nom, but fixed
according to https://github.com/Geal/nom/pull/464 (and adapted to work
outside the nom crate, and renamed my_many_till to show that it is a
local copy.

If the nom PR is accepted, the local copy of may_till should be removed.