Fix some clippy issues.
Mainly, "ascii content".as_bytes() is replaced with b"ascii content",
both in the source of ructe and in the generated code for templates.

A very similar change was suggested by @kornelski in #22, but then I
didn't know that clippy complained about it, and I thought the change
was not necessary.