An expression starting with paren ends on close.
BREAKING CHANGE.  If an expession starts with an open parenthesis, the
expression ends on the closing parenthesis (and that outer parenthesis
is not part of the expression.

This fixes cases where an expression is immediately followed by
something that could be part of the expression.

Before this change, calling a function on the result of some
subexpression could be written as `@(a - b).abs()`.  After this
change, that should be changed to `@((a - b).abs())` unless the intent
is to have the result of (a - b) followed by the template string
`.abs()`.

Fixes #64.