Don't eat whitespace after a for loop.
Fixes #68.

Whitespace was ignored after the closing brace of a for loop, which
was inconsistent with other blocks (where whitespace after the block
remains).

I might have introduced this bug intentionally way back, just to get
some less "junk" whitespace in my output. But that will have to be
solved in a more predictable and deterministic way at some later time.