Avoid modifying unmodified source.
When generating source code, write to an internal buffer first, and
then write it to file only if the file does not alreay exist or the
contents differ.  I hope this can help cargo and rustc not to bother
with unmodified sources, making incremental compilation of projects
that use ructe faster.